feat: complete security

+++ user-service/src/main/java/cuit9622/dms/user/mapper/UserMapper.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Mapper
 public interface UserMapper extends BaseMapper<User> {
-    List<String> getAuthorities(Integer userId);
+    List<String> getAuthorities(Long userId);
 
     @Select("select * from sys_user where username = #{username}")
     User getUserByUsername(String username);
